+++ vantage-aws/src/models/logs/group.rs
@@ -39,12 +39,15 @@ pub struct LogGroup {
 /// # Ok(()) }
 /// ```
 pub fn groups_table(aws: AwsAccount) -> Table<AwsAccount, LogGroup> {
-    Table::new("json1/logGroups:logs/Logs_20140328.DescribeLogGroups", aws)
-        .with_id_column("logGroupName")
-        .with_column_of::<i64>("creationTime")
-        .with_title_column_of::<i64>("storedBytes")
-        .with_many("events", "logGroupName", events_table)
-        .with_many("streams", "logGroupName", streams_table)
+    Table::new(
+        "json1/logGroups@nextToken:logs/Logs_20140328.DescribeLogGroups",
+        aws,
+    )
+    .with_id_column("logGroupName")
+    .with_column_of::<i64>("creationTime")
+    .with_title_column_of::<i64>("storedBytes")
+    .with_many("events", "logGroupName", events_table)
+    .with_many("streams", "logGroupName", streams_table)
 }
 
 impl LogGroup {